HPUIの追加及び残機処理のスクリプト変更
