+++ Assets/Prefab/HPUI.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a5e646538e6c9e8409b23ae882407482
+guid: a86eefe4f7586b641aeceb3127730cad
 PrefabImporter:
   externalObjects: {}
   userData:
