+++ Assets/Script/PlayerHP.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 65d20e235a8a1bb46a702f98b9832ac3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
